requirements.txt
face_recognition
streamlit
Pillow
numpy